# backend/docs/swagger.yaml
basePath: /
definitions:
...
  consts.NodeAccessPerm:
    enum:
    - open
    - partial
    - closed
    type: string
    x-enum-comments:
      NodeAccessPermClosed: 完全禁止
      NodeAccessPermOpen: 完全开放
      NodeAccessPermPartial: 部分开放
    x-enum-descriptions:
    - 完全开放
    - 部分开放
    - 完全禁止
    x-enum-varnames:
    - NodeAccessPermOpen
    - NodeAccessPermPartial
    - NodeAccessPermClosed
  consts.NodePermName:
    enum:
    - visible
    - visitable
    - answerable
    type: string
    x-enum-comments:
      NodePermNameAnswerable: 可被问答
      NodePermNameVisible: 导航内可见
      NodePermNameVisitable: 可被访问
    x-enum-descriptions:
    - 导航内可见
    - 可被访问
    - 可被问答
    x-enum-varnames:
    - NodePermNameVisible
    - NodePermNameVisitable
    - NodePermNameAnswerable
  consts.NodeRagInfoStatus:
    enum:
    - BASIC_PENDING
    - BASIC_RUNNING
    - BASIC_FAILED
    - BASIC_SUCCEEDED
    - ENHANCE_PENDING
    - ENHANCE_RUNNING
    - ENHANCE_FAILED
    - ENHANCE_SUCCEEDED
    type: string
    x-enum-comments:
      NodeRagStatusBasicFailed: 基础处理失败
      NodeRagStatusBasicPending: 等待基础处理
      NodeRagStatusBasicRunning: 正在进行基础处理（文本分割、向量化等）
      NodeRagStatusBasicSucceeded: 基础处理成功
      NodeRagStatusEnhanceFailed: 增强处理失败
      NodeRagStatusEnhancePending: 基础处理完成，等待增强处理
      NodeRagStatusEnhanceRunning: 正在进行增强处理（关键词提取等）
      NodeRagStatusEnhanceSucceeded: 增强处理成功
    x-enum-descriptions:
    - 等待基础处理
    - 正在进行基础处理（文本分割、向量化等）
    - 基础处理失败
    - 基础处理成功
    - 基础处理完成，等待增强处理
    - 正在进行增强处理（关键词提取等）
    - 增强处理失败
    - 增强处理成功
    x-enum-varnames:
    - NodeRagStatusBasicPending
    - NodeRagStatusBasicRunning
    - NodeRagStatusBasicFailed
    - NodeRagStatusBasicSucceeded
    - NodeRagStatusEnhancePending
    - NodeRagStatusEnhanceRunning
    - NodeRagStatusEnhanceFailed
    - NodeRagStatusEnhanceSucceeded
...
  domain.BatchMoveReq:
    properties:
      ids:
        items:
          type: string
        type: array
      kb_id:
        type: string
      parent_id:
        type: string
    required:
    - ids
    - kb_id
    type: object
...
  domain.CreateNodeReq:
    properties:
      content:
        type: string
      content_type:
        type: string
      emoji:
        type: string
      kb_id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      position:
        type: number
      summary:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domain.NodeType'
        enum:
        - 1
        - 2
    required:
    - kb_id
    - name
    - type
    type: object
...
  domain.MoveNodeReq:
    properties:
      id:
        type: string
      kb_id:
        type: string
      next_id:
        type: string
      parent_id:
        type: string
      prev_id:
        type: string
    required:
    - id
    - kb_id
    type: object
  domain.NodeActionReq:
    properties:
      action:
        enum:
        - delete
        type: string
      ids:
        items:
          type: string
        type: array
      kb_id:
        type: string
    required:
    - action
    - ids
    - kb_id
    type: object
...
  domain.NodeGroupDetail:
    properties:
      auth_group_id:
        type: integer
      auth_ids:
        items:
          type: integer
        type: array
      kb_id:
        type: string
      name:
        type: string
      node_id:
        type: string
      perm:
        $ref: '#/definitions/consts.NodePermName'
    type: object
  domain.NodeListItemResp:
    properties:
      content_type:
        type: string
      created_at:
        type: string
      creator:
        type: string
      creator_id:
        type: string
      editor:
        type: string
      editor_id:
        type: string
      emoji:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      position:
        type: number
      publisher_id:
        type: string
      rag_info:
        $ref: '#/definitions/domain.RagInfo'
      status:
        $ref: '#/definitions/domain.NodeStatus'
      summary:
        type: string
      type:
        $ref: '#/definitions/domain.NodeType'
      updated_at:
        type: string
    type: object
  domain.NodeMeta:
    properties:
      content_type:
        type: string
      emoji:
        type: string
      summary:
        type: string
    type: object
  domain.NodePermissions:
    properties:
      answerable:
        allOf:
        - $ref: '#/definitions/consts.NodeAccessPerm'
        description: 可被问答
      visible:
        allOf:
        - $ref: '#/definitions/consts.NodeAccessPerm'
        description: 导航内可见
      visitable:
        allOf:
        - $ref: '#/definitions/consts.NodeAccessPerm'
        description: 可被访问
    type: object
  domain.NodeStatus:
    enum:
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - NodeStatusDraft
    - NodeStatusReleased
  domain.NodeSummaryReq:
    properties:
      ids:
        items:
          type: string
        type: array
      kb_id:
        type: string
    required:
    - ids
    - kb_id
    type: object
  domain.NodeType:
    enum:
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - NodeTypeFolder
    - NodeTypeDocument
...
  domain.PWResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.RagInfo:
    properties:
      message:
        type: string
      status:
        $ref: '#/definitions/consts.NodeRagInfoStatus'
    type: object
  domain.RecommendNodeListResp:
    properties:
      emoji:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      position:
        type: number
      recommend_nodes:
        items:
          $ref: '#/definitions/domain.RecommendNodeListResp'
        type: array
      summary:
        type: string
      type:
        $ref: '#/definitions/domain.NodeType'
    type: object
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.ShareNodeDetailItem:
    properties:
      children:
        items:
          $ref: '#/definitions/domain.ShareNodeDetailItem'
        type: array
      emoji:
        type: string
      id:
        type: string
      meta:
        $ref: '#/definitions/domain.NodeMeta'
      name:
        type: string
      parent_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      position:
        type: number
      type:
        $ref: '#/definitions/domain.NodeType'
      updated_at:
        type: string
    type: object
...
  domain.UpdateNodeReq:
    properties:
      content:
        type: string
      content_type:
        type: string
      emoji:
        type: string
      id:
        type: string
      kb_id:
        type: string
      name:
        type: string
      position:
        type: number
      summary:
        type: string
    required:
    - id
    - kb_id
    type: object
...
  v1.NodeDetailResp:
    properties:
      content:
        type: string
      created_at:
        type: string
      creator_account:
        type: string
      creator_id:
        type: string
      editor_account:
        type: string
      editor_id:
        type: string
      id:
        type: string
      kb_id:
        type: string
      meta:
        $ref: '#/definitions/domain.NodeMeta'
      name:
        type: string
      parent_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      publisher_account:
        type: string
      publisher_id:
        type: string
      pv:
        type: integer
      status:
        $ref: '#/definitions/domain.NodeStatus'
      type:
        $ref: '#/definitions/domain.NodeType'
      updated_at:
        type: string
    type: object
  v1.NodePermissionEditReq:
    properties:
      answerable_groups:
        description: 可被问答
        items:
          type: integer
        type: array
      ids:
        items:
          type: string
        type: array
      kb_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      visible_groups:
        description: 导航内可见
        items:
          type: integer
        type: array
      visitable_groups:
        description: 可被访问
        items:
          type: integer
        type: array
    required:
    - ids
    - kb_id
    type: object
  v1.NodePermissionEditResp:
    type: object
  v1.NodePermissionResp:
    properties:
      answerable_groups:
        description: 可被问答
        items:
          $ref: '#/definitions/domain.NodeGroupDetail'
        type: array
      id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      visible_groups:
        description: 导航内可见
        items:
          $ref: '#/definitions/domain.NodeGroupDetail'
        type: array
      visitable_groups:
        description: 可被访问
        items:
          $ref: '#/definitions/domain.NodeGroupDetail'
        type: array
    type: object
  v1.NodeRestudyReq:
    properties:
      kb_id:
        type: string
      node_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - kb_id
    - node_ids
    type: object
  v1.NodeRestudyResp:
    type: object
...
  v1.ShareNodeDetailResp:
    properties:
      content:
        type: string
      created_at:
        type: string
      creator_account:
        type: string
      creator_id:
        type: string
      editor_account:
        type: string
      editor_id:
        type: string
      id:
        type: string
      kb_id:
        type: string
      list:
        items:
          $ref: '#/definitions/domain.ShareNodeDetailItem'
        type: array
      meta:
        $ref: '#/definitions/domain.NodeMeta'
      name:
        type: string
      parent_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      publisher_account:
        type: string
      publisher_id:
        type: string
      pv:
        type: integer
      status:
        $ref: '#/definitions/domain.NodeStatus'
      type:
        $ref: '#/definitions/domain.NodeType'
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: panda-wiki API documentation
  title: panda-wiki API
  version: "1.0"
paths:
...
  /api/v1/node:
    post:
      consumes:
      - application/json
      description: Create Node
      parameters:
      - description: Node
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateNodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - bearerAuth: []
      summary: Create Node
      tags:
      - node
  /api/v1/node/action:
    post:
      consumes:
      - application/json
      description: Node Action
      parameters:
      - description: Action
        in: body
        name: action
        required: true
        schema:
          $ref: '#/definitions/domain.NodeActionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - bearerAuth: []
      summary: Node Action
      tags:
      - node
  /api/v1/node/batch_move:
    post:
      consumes:
      - application/json
      description: Batch Move Node
      parameters:
      - description: Batch Move Node
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.BatchMoveReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Batch Move Node
      tags:
      - node
  /api/v1/node/detail:
    get:
      consumes:
      - application/json
      description: Get Node Detail
      parameters:
      - in: query
        name: format
        type: string
      - in: query
        name: id
        required: true
        type: string
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.NodeDetailResp'
              type: object
      security:
      - bearerAuth: []
      summary: Get Node Detail
      tags:
      - node
    put:
      consumes:
      - application/json
      description: Update Node Detail
      parameters:
      - description: Node
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateNodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Update Node Detail
      tags:
      - node
  /api/v1/node/list:
    get:
      consumes:
      - application/json
      description: Get Node List
      parameters:
      - in: query
        name: kb_id
        required: true
        type: string
      - in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.NodeListItemResp'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: Get Node List
      tags:
      - node
  /api/v1/node/move:
    post:
      consumes:
      - application/json
      description: Move Node
      parameters:
      - description: Move Node
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.MoveNodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Move Node
      tags:
      - node
  /api/v1/node/permission:
    get:
      consumes:
      - application/json
      description: 文档授权信息获取
      operationId: v1-NodePermission
      parameters:
      - in: query
        name: id
        required: true
        type: string
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.NodePermissionResp'
              type: object
      security:
      - bearerAuth: []
      summary: 文档授权信息获取
      tags:
      - NodePermission
  /api/v1/node/permission/edit:
    patch:
      consumes:
      - application/json
      description: 文档授权信息更新
      operationId: v1-NodePermissionEdit
      parameters:
      - description: para
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/v1.NodePermissionEditReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.NodePermissionEditResp'
              type: object
      security:
      - bearerAuth: []
      summary: 文档授权信息更新
      tags:
      - NodePermission
  /api/v1/node/recommend_nodes:
    get:
      consumes:
      - application/json
      description: Recommend Nodes
      parameters:
      - in: query
        name: kb_id
        required: true
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: node_ids
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.RecommendNodeListResp'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: Recommend Nodes
      tags:
      - node
  /api/v1/node/restudy:
    post:
      consumes:
      - application/json
      description: 文档重新学习
      operationId: v1-NodeRestudy
      parameters:
      - description: para
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/v1.NodeRestudyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.NodeRestudyResp'
              type: object
      security:
      - bearerAuth: []
      summary: 文档重新学习
      tags:
      - Node
  /api/v1/node/summary:
    post:
      consumes:
      - application/json
      description: Summary Node
      parameters:
      - description: Summary Node
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.NodeSummaryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Summary Node
      tags:
      - node
...
  /share/v1/node/detail:
    get:
      consumes:
      - application/json
      description: GetNodeDetail
      parameters:
      - description: kb id
        in: header
        name: X-KB-ID
        required: true
        type: string
      - description: node id
        in: query
        name: id
        required: true
        type: string
      - description: format
        in: query
        name: format
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.ShareNodeDetailResp'
              type: object
      summary: GetNodeDetail
      tags:
      - share_node
  /share/v1/node/list:
    get:
      consumes:
      - application/json
      description: GetNodeList
      parameters:
      - description: kb id
        in: header
        name: X-KB-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: GetNodeList
      tags:
      - share_node
...
securityDefinitions:
  bearerAuth:
    description: Type "Bearer" + a space + your token to authorize
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"